requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# app.py: 7
Flask == 1.0.2

# app.py: 8
Flask_Cors == 3.0.8

# models\cnn.py: 6
Pillow == 5.4.1

# models\cnn.py: 5
numpy == 1.16.2

# models\cnn.py: 11
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl

# models\cnn.py: 10
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp37-cp37m-linux_x86_64.whl
